I created a [IConnectable] interface that will be inhereted by the [DatabaseConnector] to implement basic functionality needed for the database connection  and query execution.
Added a 'ConnenctionState' to the [DatabaseConnector] so it keep tracks of the connection.

Created the private members for the connection and command. These are polymorphic assign. It is able to hold different types of database connector and it will execute the correct connection base on the type at runtime. Only one connector 'MySqlConnector' is being implemented.

Created a enum type for the basic DML SQL commands that I will be using. Example ['Select', 'Update', 'Delete'].

Created a dictionary that reference those enums to a string representation of those action for readability.[Improvent] use a static class as a enum similar to the [Provider] to allocate those value on the stack instead of the heap.

Created a [Provider.MySql] that hold the name of the 'MySQLClient' provider.

Implemented the Open, Close, Command(str), Execute() methods for [DatabaseConnector]. The Command(str) and Execute() will be run as a pipeline command. The Execute() method inspect the string and redirect to the right function base on found command substring.

Created some eventd handlers, eventd triggers, events invokationd for when connection status change. I will potential implement one for exceptions trigger.

Created a struct [Data] that holds the information that the 'Readers' SQL object reads back from the executed command. Such informaton as the count, values, and datatypes.

The query has been tested multiple times succesfully, the data has been refactored.