Many refactoring changes including:
- Created a Class Library (.dll) for the .NetFramework so it can be exchange from a console app to a Windows' Form app.

  Seperated the Model (Logic) from the (.exe) console app where only should hold control (Controller) and prensentation (View) data.

Created a [IDatabaseConnector] which establish the properties and signatures that the data base connector object should implement such as open connection, close connection, create command, execute.

Created the [MySqlConnector] class that implements the [IDatabaseConnector] interface. This is use to fetch and update data that has a MySql connection.

Created a enum/static class that holds specific structure and literal values in one central place for the application. One central location for updates and changes.

Created a [DatabaseDataTable] class that uses a return [IReader] interface to construct a [DataTable] from the fetched data returned from the database. Although, not implemented yet this class will be use to create the [DataTable] object that should be updated on the database object.

Created two instances classes/factory class [DatabaseInstance] & [DataInstance] where the first one, return newly created database instance centrallize in one location and it will be mostly use in a [IDatabaseConnector] object and the second is returns newly created objects that will be use in the [DatabaseDataTable] class.