Moved plenty of classes here to a [SchedulingLibrary] .dll to separate logic from the control and ui/output implementation
Created a [Application] class that will be instatiated when the [Program] class execute. It will have the execution function that is being tested. Different ones in the future if a different type of test is needed.

Created the [Controller] class that will be use to map the users inputs once the application is ready for user inputs. This class has not yet been implemented.

Created the [ConsoleOutput] class that will format specific output for the console. In this case it has a function that displays the [DataTable] in a table format. It has been tested and it displayed the data table that had data stored from the reader that returned from the database.