Name Changes: [DatabaseDataTable] => [DbDataSet], [DatabaseStructure] => [DbStructure], [IDatabaseConnector] => [IDbConnector], [DatabaseInstace] => [DbInstance], [Structures] => [Structure].
Folder Changes: New (factory) folder to store the [DbInstance] and [DataInstance] factory classes.

Redesigned the [DbDataSet] to map and fill and create cronstraints of all the tables of the the provided database.

Created a Abstract class [DbSchema] that has the basic requirements needed for the [DbDataSet] such as database name, table names, primary key name(s) (Composite Key), foreign key names as well as the primary/foreign keys table names relationship, that will match the sequence of the foreign keys implying which table is the foreign key the primary. The [DbSchema] abstract class allows in the future to develop another class with different database information without breaking the [DbDataSet] implementation.

Removed other properties such as DbCommand and DataReader from the [IDbConnector]. Only kept the IDbDataAdapter] which is the one currently needed.

Readjusted the DbDataAdapter to create an DbDataAdapter instance if it is null when you access it.