Changed the naming convention for classes, namespaces and methods to reflect the current C# .Net suggested standard.
Created different namespaces and folders to compartilize more the behavior.

Changed the [Program] class to start the [Application] class with the database provider name in its constructor and changed the Run function to take no paramaters after.

Changed the [ConsoleOutput] ShowTable function to private and created the same function but with a different signature (overloaded) to take a [DataTableColletion] parameter that iterates through all the tables and call the internal ShowTable function, displaying each table individually.

Created a [IDbConfiguration] interface that establish a few properties need to identify the parameters to create a database connection.

Created the [MySqlConfiguration] that implements the [IDbConfiguration] and uses the establish properties to set the connection string and connection type.

Created the [AppFactory] class that instantiated the configuration, database connector, and database dataset objects use in the application. They are stored on the [AppState].

Created a [AppState] class that will hold different components and objects representing the state of the application.

Moved the testing functions to a new [Tester] static class to invoke each test. Each test takes a [AppState] as parameters.

Updated the [Controller] => [AppController] class.