Removed listed files on the 'gitignore'.
Moved the Config folder with its classes [IDbConfiguration, MySqlConfiguration] from the [Scheduling Console App] > [Scheduling Library].

Changed one of the property names from 'TableNames' to 'TableNamesIndented' since there is a class that will have also the table names.

Changed the [ClientScheduleDbSchema] and created multiple classes that describe the column names in all the tables of the client_schedule database.

Reformated the [DbDataSet] to reflect the new changes on the [ClientScheduleDbSchema]. Implementation is the same.

Created a [ClientScheduleRecord] class that holds multiple classes that represent the record/data of a table. It will be used to store the UI component information.

Modified the [Application] class to provide the provider's name and database's name. Also, I removed some of the instantiation of the [IDbConnector] and [DbDataSet] to place on the [AppState] class.

Modified the [AppData] class to hold the DbRecords in the [AppState] to hold it based on the database name. This will ensure the AppState has the right database records.

Modified the [Tester] class that tested a couple failing and passing scenarios on the 'AuthenticationTest' function.

Modified the [AppFactory] to create the [DbSchema] and the [AppData] and more.

Modified the [AppState] class to use the [AppFactory] for instantiation of its new members. Changed the parameterized constructor to receive a [IDConfiguration] and database name.