Changed the class name from [MySqlConnector] to more specific [DbConnector] class.
Removed the [IDbConnector] interface. It is not implmented anymore.

Refactored the [DbConnector] and made it more specific and to fit the mapping, update insert and delete processing using the DataSet from the [DbDataSet] class. It uses dbfactory instance to create its class components.

Refactored the [DbDataSet] and seperated all the database components actions into the [DbConnector] that is aggregated in the [DbDataSet]. Now the [DbDataset] does not have database component statements anymore (OOP). This class adapts to different databases provided base on the [IDbConfig] interface that hass information about the connection string.

Created the [DBFactory] to instantiated the 'DBProviderFactory'.

Removed the in (dereferencing) on the functions parameters.