Created the [LoginForm] that holds the [LoginControl].
Modified the [LoginControl] so it display tooltips when the textboxes are empty. Rearranged the canvas and components sizes and points/positions.

Added the warning icon on the tooltips and implemented the event handlder behavior when the textboxes looses focus. If the string are empty it will display a message based on the [LoginControlValidator].