Created a [LoadingControl] that handles loading animation that uses a background work from the [AppForm] that displays and hides depending on the loading function time and action. Added multiple spinner image resources to the application that asynchronously iteratios through the animation.
The [LoginControl] has a [AppState] prop to be assigned and use on. It uses the [AppState] StaticValidateAppStateForNull to display the appropiate message when error. It also has the tooltips when the value are missing on the textfield.

Created an Abort function in the [Program] static class use for unrecoverble exceptions.

Updated the [LoginForm] to [LoginCanvasControl] and it is appending to the main [AppForm]. It displays a message box when the programs runs into a fatal exception. It also have a nice background image that blend with the other components.

Added a UI-StyleUtility folder that has [CenterWinDialog] to center the dialogbox in the center of the form when display.

Created the [AppForm] that handles the UIState/AppState interaction with the database with a background worker component asynchronously. It stores a exception and has an exception event handler that will trigger if any fatal exception stops the application from the back end. Makes the controls color change to transparent to match the background. Shows and hides control based on the state as needed.

The applicaton is able to update the a textbox if the user was able to sign or not. This will be later replace with the new Control.