Created a [NetworkDataProvider] static class that uses the 'HttpClient' and a API key from http://ipconfig.me to get the user's location. It also have asynchronous to get the user's location and timezone based on the user's machine. Have a string that suppose to store the API key, it has been removed.
On the [ClientScheduleDbSchema], i added five functions Get<tableName>ColumnNames that return a string array all all the column names of the specific table.

Added a 'TimeZone' and 'Location' property to the [AppState].

Implemented the Insert function in the [DbConnector] to commit the data to the database.

[AppController] AddCustomerRecord function has been implemented and succesfully executed.

[AppData] have been extensevely modified to store the create the insert statements. Created 'AddUserValAndDefaultVal' function that add all the default values as well as the user's values to different records for each table. It validates and checkes for specific ids if found to continue them on the same table sequency id. Created mulitple functions that create a insert statement string for each table.

Implemented the Insert function that gathers multiple things from the appstate to insert into the DataSet and the database. Changed the 'Mapped' private member to pulic property so it can be use by other classes to validate.

Created a [TimeConverter] class that changes the UTC database time to the user's local time and vice versa. Has not been tested yet.

Started implemented  the 'CheckTimeIsOnBusinnesHours' function [Validator] class.